清除不需要的第三方框架

+++ OAManage/Classes/触控/viewController/RSStockViewController.m
@@ -8,16 +8,12 @@
 
 #import "RSStockViewController.h"
 #import "RSBLStockCell.h"
-
 #import "RSSLStockHeaderView.h"
 #import "RSSLStockFootView.h"
 #import "RSSLStockCell.h"
 #import "RSTouchAlertView.h"
-
-
 //荒料的模型
 #import "RSBMStockModel.h"
-
 //大板的模型
 #import "RSSlStockModel.h"
 #import "RSSLPieceModel.h"
@@ -35,19 +31,17 @@
 @property (nonatomic,strong)RSTouchAlertView * touchAlertview;
 
 @property (nonatomic,assign)NSInteger pageNum;
-
 //荒料号
 @property (nonatomic,strong)NSString * blockNo;
 //物料名称
 @property (nonatomic,strong)NSString * mtlName;
 
-
 @property (nonatomic,strong)UITableView * contentTableview;
-
 
 @end
 
 @implementation RSStockViewController
+
 - (RSTouchAlertView *)touchAlertview{
     if (!_touchAlertview) {
         _touchAlertview = [[RSTouchAlertView alloc]initWithFrame:CGRectMake(33, SCH/2 - 180, SCW - 66, 290)];
@@ -71,16 +65,12 @@
 //    return _contentTableview;
 //}
 
-
-
-
 - (NSMutableArray *)choosingArray{
     if (!_choosingArray) {
         _choosingArray = [NSMutableArray array];
     }
     return _choosingArray;
 }
-
 
 - (NSMutableArray *)expendArray{
     if (!_expendArray) {
@@ -103,8 +93,6 @@
     return _selectdeContentArray;
 }
 
-
-
 - (NSMutableArray *)selectionArray{
     if (!_selectionArray) {
         _selectionArray = [NSMutableArray array];
@@ -112,32 +100,22 @@
     return _selectionArray;
 }
 
-
-
 static NSString * SLSTOCKHEADERVIEWID = @"SLSTOCKHEADERVIEWID";
 static NSString * SLSTOCKFOOTVIEWID = @"SLSTOCKFOOTVIEWID";
 - (void)viewDidLoad {
     [super viewDidLoad];
-    
 //    self.title = @"荒料库存";
     self.emptyView.hidden = YES;
-    
 //    self.tableview.backgroundColor = [UIColor colorWithHexColorStr:@"#f5f5f5"];
     self.contentTableview.backgroundColor = [UIColor colorWithHexColorStr:@"#f5f5f5"];
-    
    // self.tableview.contentInset = UIEdgeInsetsMake(0, 0, 40, 0);
-    
     self.contentTableview.contentInset = UIEdgeInsetsMake(0, 0, 100, 0);
-    
-    
     [self.tableview removeFromSuperview];
-    
      if (@available(iOS 11.0, *)) {
-            [[UIScrollView appearance] setContentInsetAdjustmentBehavior:UIScrollViewContentInsetAdjustmentAutomatic];
+        [[UIScrollView appearance] setContentInsetAdjustmentBehavior:UIScrollViewContentInsetAdjustmentAutomatic];
     }else{
         self.automaticallyAdjustsScrollViewInsets = NO;
     }
-        
     UITableView * contentTableview = [[UITableView alloc]initWithFrame:CGRectMake(0, CGRectGetMaxY(self.navigationController.navigationBar.frame), SCW, SCH - CGRectGetMaxY(self.navigationController.navigationBar.frame)) style:UITableViewStyleGrouped];
     contentTableview.separatorStyle = UITableViewCellSeparatorStyleNone;
     contentTableview.delegate = self;
@@ -147,12 +125,8 @@ static NSString * SLSTOCKFOOTVIEWID = @"SLSTOCKFOOTVIEWID";
     contentTableview.estimatedSectionFooterHeight = 0;
     _contentTableview = contentTableview;
     [self.view addSubview:self.contentTableview];
-    
-    
     self.blockNo = @"";
     self.mtlName = @"";
-    
-    
     UIButton * screenBtn = [[UIButton alloc]initWithFrame:CGRectMake(0, 0, 50, 50)];
     [screenBtn setTitle:@"筛选" forState:UIControlStateNormal];
     [screenBtn setTitleColor:[UIColor colorWithHexColorStr:@"#333333"] forState:UIControlStateNormal];
@@ -160,13 +134,10 @@ static NSString * SLSTOCKFOOTVIEWID = @"SLSTOCKFOOTVIEWID";
     [screenBtn addTarget:self action:@selector(screenAction:) forControlEvents:UIControlEventTouchUpInside];
     UIBarButtonItem * testitem = [[UIBarButtonItem alloc]initWithCustomView:screenBtn];
     self.navigationItem.rightBarButtonItem = testitem;
-    
     self.pageNum = 1;
     self.contentTableview.mj_header = [MJChiBaoZiHeader headerWithRefreshingTarget:self refreshingAction:@selector(reloadAuditedNewData)];
     self.contentTableview.mj_footer = [MJChiBaoZiFooter footerWithRefreshingTarget:self refreshingAction:@selector(reloadAuditedMoreNewData)];
     [self.contentTableview.mj_header beginRefreshing];
-    
-
     [self setUIBottomView];
 }
 
@@ -179,16 +150,12 @@ static NSString * SLSTOCKFOOTVIEWID = @"SLSTOCKFOOTVIEWID";
     [self reloadStockData];
 }
 
-
-
-
 - (void)reloadStockData{
     NSUserDefaults * user = [NSUserDefaults standardUserDefaults];
     NSString * aes = [user objectForKey:@"AES"];
     NSString *const kInitVector = @"16-Bytes--String";
     NetworkTool * network = [[NetworkTool alloc]init];
     NSNumber * number = [NSNumber numberWithInteger:self.pageNum];
-    
     NSString * type = [NSString string];
     if ([self.selectType isEqualToString:@"huangliao"]) {
         type = @"BM";
@@ -205,7 +172,6 @@ static NSString * SLSTOCKFOOTVIEWID = @"SLSTOCKFOOTVIEWID";
         NSNumber * storeAreaId = [NSNumber numberWithInteger:self.storeAreamodel.storeAreaId];
         notice = URL_YIGODATA_SLSTOCK(number, @(10), self.mtlName, self.blockNo, deaid, type, whsId, storeAreaId);
     }
-    
     NSString * aes2 = [FSAES128 encryptAES:notice key:aes andKInItVector:kInitVector];
     NSString * canshu = URL_YIGODATA_NOTICE(self.usermodel.appLoginToken, aes2);
     NSString * sopaStr = URL_YIGODATA_IOS(URL_WORKFLOWWEBSERVICE, URL_STOCK, canshu);
@@ -217,13 +183,11 @@ static NSString * SLSTOCKFOOTVIEWID = @"SLSTOCKFOOTVIEWID";
                  [self.choosingArray removeAllObjects];
                  [self.choosingArray addObjectsFromArray:array];
                  self.pageNum = 2;
-                 
                  [self.contentTableview.mj_header endRefreshing];
             }else{
                  NSArray * array1 = array;
                  [self.choosingArray addObjectsFromArray:array1];
                  self.pageNum++;
-                 
                  [self.contentTableview.mj_footer endRefreshing];
             }
             [self.contentTableview reloadData];
@@ -242,7 +206,6 @@ static NSString * SLSTOCKFOOTVIEWID = @"SLSTOCKFOOTVIEWID";
                 [self.contentTableview.mj_header endRefreshing];
                 [self.contentTableview.mj_footer endRefreshing];
         };
-        
     }else{
         network.successReload = ^(NSDictionary *dict) {
             NSMutableArray * array = [NSMutableArray array];
@@ -257,9 +220,7 @@ static NSString * SLSTOCKFOOTVIEWID = @"SLSTOCKFOOTVIEWID";
                     slstockmodel.msid = [[array objectAtIndex:i]objectForKey:@"msid"];
                     slstockmodel.mtlName = [[array objectAtIndex:i]objectForKey:@"mtlName"];
                     slstockmodel.turnsNo = [[array objectAtIndex:i]objectForKey:@"turnsNo"];
-                    
                     tempArray = [[array objectAtIndex:i]objectForKey:@"piece"];
-                    
                     for (int j = 0; j < tempArray.count; j++) {
                         RSSLPieceModel * slpiecemodel = [[RSSLPieceModel alloc]init];
                         slpiecemodel.area = [[tempArray objectAtIndex:j]objectForKey:@"area"];
@@ -276,7 +237,6 @@ static NSString * SLSTOCKFOOTVIEWID = @"SLSTOCKFOOTVIEWID";
                     [self.choosingArray addObject:slstockmodel];
                 }
                 self.pageNum = 2;
-               
                 [self.contentTableview.mj_header endRefreshing];
             }else{
                 NSMutableArray * tempArray = [NSMutableArray array];
@@ -309,7 +269,6 @@ static NSString * SLSTOCKFOOTVIEWID = @"SLSTOCKFOOTVIEWID";
                 //[self.contentTableview reloadData];
                 [self.contentTableview.mj_footer endRefreshing];
             }
-            
              [self.contentTableview reloadData];
 //            if (self.choosingArray.count > 0 ) {
 //                self.emptyView.hidden = YES;
@@ -329,20 +288,9 @@ static NSString * SLSTOCKFOOTVIEWID = @"SLSTOCKFOOTVIEWID";
     }
 }
 
-
-
-
-
-
-
-
-
-
-
 - (void)screenAction:(UIButton *)screenBtn{
      [self.touchAlertview showView];
 }
-
 
 - (NSInteger)numberOfSectionsInTableView:(UITableView *)tableView{
     if (tableView == self.contentTableview) {
@@ -353,7 +301,6 @@ static NSString * SLSTOCKFOOTVIEWID = @"SLSTOCKFOOTVIEWID";
         }
     }
     return 0;
-    
 }
 
 - (NSInteger)tableView:(UITableView *)tableView numberOfRowsInSection:(NSInteger)section{
@@ -370,11 +317,9 @@ static NSString * SLSTOCKFOOTVIEWID = @"SLSTOCKFOOTVIEWID";
           }
     }
     return 0;
-  
 }
 
 - (CGFloat)tableView:(UITableView *)tableView heightForHeaderInSection:(NSInteger)section{
-    
     if (tableView == self.contentTableview) {
         if ([self.selectType isEqualToString:@"daban"]) {
               return 148;
@@ -382,12 +327,8 @@ static NSString * SLSTOCKFOOTVIEWID = @"SLSTOCKFOOTVIEWID";
          }else{
               return 0.001;
          }
-         
-        
-        
     }
     return 0;
- 
 }
 
 - (CGFloat)tableView:(UITableView *)tableView heightForFooterInSection:(NSInteger)section{
@@ -397,12 +338,9 @@ static NSString * SLSTOCKFOOTVIEWID = @"SLSTOCKFOOTVIEWID";
         }else{
             return 0.001;
         }
-        
     }
     return 0;
-    
 }
-
 
 - (CGFloat)tableView:(UITableView *)tableView heightForRowAtIndexPath:(NSIndexPath *)indexPath{
     if (tableView == self.contentTableview) {
@@ -413,9 +351,7 @@ static NSString * SLSTOCKFOOTVIEWID = @"SLSTOCKFOOTVIEWID";
            }
     }
     return 0;
-   
 }
-
 
 - (UIView *)tableView:(UITableView *)tableView viewForHeaderInSection:(NSInteger)section{
     if (tableView == self.contentTableview) {
@@ -424,9 +360,7 @@ static NSString * SLSTOCKFOOTVIEWID = @"SLSTOCKFOOTVIEWID";
             slstockHeaderview.selectBtn.tag = section;
             slstockHeaderview.tag = section;
             slstockHeaderview.tap.view.tag = section;
-            
             RSSlStockModel * slstockmodel = self.choosingArray[section];
-            
             slstockHeaderview.slstockmodel = slstockmodel;
             if (self.selectdeContentArray.count > 0) {
                 for (int j = 0; j < self.selectdeContentArray.count; j++) {
@@ -469,7 +403,6 @@ static NSString * SLSTOCKFOOTVIEWID = @"SLSTOCKFOOTVIEWID";
             }
             [slstockHeaderview.selectBtn addTarget:self action:@selector(dabanHeaderButtonOnClickAction:) forControlEvents:UIControlEventTouchUpInside];
             [slstockHeaderview.tap addTarget:self action:@selector(dabanHeaderTapAction:)];
-            
             return slstockHeaderview;
         }else{
             return nil;
@@ -477,7 +410,6 @@ static NSString * SLSTOCKFOOTVIEWID = @"SLSTOCKFOOTVIEWID";
     }
     return nil;
 }
-
 
 - (UIView *)tableView:(UITableView *)tableView viewForFooterInSection:(NSInteger)section{
     if (tableView == self.contentTableview) {
@@ -487,16 +419,11 @@ static NSString * SLSTOCKFOOTVIEWID = @"SLSTOCKFOOTVIEWID";
         }else{
             return nil;
         }
-        
     }
     return nil;
-    
 }
 
-
-
 - (UITableViewCell *)tableView:(UITableView *)tableView cellForRowAtIndexPath:(NSIndexPath *)indexPath{
-    
     if (tableView == self.contentTableview) {
         if ([self.selectType isEqualToString:@"huangliao"]) {
             static NSString * STOCKCELLID = @"STOCKCELLID";
@@ -507,7 +434,6 @@ static NSString * SLSTOCKFOOTVIEWID = @"SLSTOCKFOOTVIEWID";
             RSBMStockModel * bmstockmodel = self.choosingArray[indexPath.row];
             cell.bmstockmodel = bmstockmodel;
             cell.selectBtn.tag = indexPath.row;
-            
             if (self.selectionArray.count > 0) {
                  for (int i = 0; i < self.selectionArray.count; i++) {
                  RSBMStockModel * bmstockmodel1 = self.selectionArray[i];
@@ -525,7 +451,6 @@ static NSString * SLSTOCKFOOTVIEWID = @"SLSTOCKFOOTVIEWID";
             [cell.selectBtn addTarget:self action:@selector(huangliaoSelectAction:) forControlEvents:UIControlEventTouchUpInside];
             cell.selectionStyle = UITableViewCellSelectionStyleNone;
             return cell;
-            
         }else{
             static NSString * STOCKCELLID = @"STOCKCELLID";
             RSSLStockCell * cell = [tableView dequeueReusableCellWithIdentifier:STOCKCELLID];
@@ -535,13 +460,10 @@ static NSString * SLSTOCKFOOTVIEWID = @"SLSTOCKFOOTVIEWID";
             cell.tag = indexPath.section;
             cell.selectBtn.tag = indexPath.row;
             [cell.selectBtn addTarget:self action:@selector(dabanShowAndHideSelectContentAction:) forControlEvents:UIControlEventTouchUpInside];
-            
             RSSlStockModel * slstockmodel = self.choosingArray[indexPath.section];
             RSSLPieceModel * slpiecemodel = slstockmodel.piece[indexPath.row];
-            
-            cell.filmNumberLabel.text = [NSString stringWithFormat:@"片号%ld",indexPath.row + 1];
+            cell.filmNumberLabel.text = [NSString stringWithFormat:@"片号%ld",(long)(indexPath.row + 1)];
             cell.longDetailLabel.text = [NSString stringWithFormat:@"%0.3lf",[slpiecemodel.area doubleValue]];
-            
             if (self.selectdeContentArray.count > 0) {
                    for (int i = 0; i < self.selectdeContentArray.count; i++) {
                        RSSLPieceModel * slpiecemodel1 = self.selectdeContentArray[i];
@@ -569,15 +491,8 @@ static NSString * SLSTOCKFOOTVIEWID = @"SLSTOCKFOOTVIEWID";
             cell.selectionStyle = UITableViewCellSelectionStyleNone;
             return cell;
         }
-        
-        
-        
-        
     }
-    
     return nil;
-    
-    
 }
 
 - (void)setUIBottomView{
@@ -716,8 +631,6 @@ static NSString * SLSTOCKFOOTVIEWID = @"SLSTOCKFOOTVIEWID";
     [self.contentTableview reloadData];
 }
 
-
-
 //rstouchAlertViewDelegate
 - (void)inputName:(NSString *)string andBlockName:(NSString *)blockName{
     self.mtlName = string;
@@ -726,6 +639,5 @@ static NSString * SLSTOCKFOOTVIEWID = @"SLSTOCKFOOTVIEWID";
     [self reloadStockData];
 
 }
-
 
 @end
